Move database initializer into db folder
