+++ forms/forms_suite_test.go
@@ -15,5 +15,5 @@ func TestForms(t *testing.T) {
 }
 
 var _ = BeforeSuite(func() {
-	bootstrap.SetUpDB()
+	bootstrap.SetUp()
 })
